Introduce store/utils.js and export all utils, types
